+++ src/main/java/seedu/address/logic/commands/TaskEditCommand.java
@@ -23,7 +23,7 @@ import seedu.address.model.datetime.DateCustom;
 import seedu.address.model.datetime.TimeCustom;
 import seedu.address.model.patient.Patient;
 import seedu.address.model.person.Person;
-import seedu.address.model.record.LinkedPatient;
+import seedu.address.model.task.LinkedPatient;
 import seedu.address.model.task.Priority;
 import seedu.address.model.task.Task;
 import seedu.address.model.task.Title;
