+++ src/main/java/seedu/address/storage/JsonAdaptedTask.java
@@ -8,7 +8,7 @@ import seedu.address.model.datetime.DateCustom;
 import seedu.address.model.datetime.TimeCustom;
 import seedu.address.model.patient.Nric;
 import seedu.address.model.person.Name;
-import seedu.address.model.record.LinkedPatient;
+import seedu.address.model.task.LinkedPatient;
 import seedu.address.model.task.Priority;
 import seedu.address.model.task.Task;
 import seedu.address.model.task.Title;
